Move functions from utils to dom-utils
