+++ packages/atomic/src/components/search/atomic-search-box/atomic-search-box.tsx
@@ -33,11 +33,8 @@ import {
 } from '../../../utils/local-storage-utils';
 import {updateBreakpoints} from '../../../utils/replace-breakpoint-utils';
 import {AriaLiveRegion} from '../../../utils/stencil-accessibility-utils';
-import {
-  isFocusingOut,
-  once,
-  spreadProperties
-} from '../../../utils/utils';
+import {isFocusingOut} from '../../../utils/dom-utils';
+import {once, spreadProperties} from '../../../utils/utils';
 import {randomID} from '../../../utils/random-utils';
 import {RedirectionPayload} from '../../common/search-box/redirection-payload';
 import {SearchBoxWrapper} from '../../common/search-box/stencil-search-box-wrapper';
